+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This document notes all API visible changes of this library.
 - Added JSON conversion support for `ResourceLocation`.
 - Added JSON conversion support for NBT and Keybind chat component.
 - Added XML documentation.
+- Added `JsonComponentSerializer` utility class that helps converting a chat component.
 
 ### Changed